refactor(llmobs): rename openai-agents-core dirs to openai-agents
Align LLMObs plugin and test directory names with the integration/plugin
ID (openai-agents) rather than the npm sub-package name (openai-agents-core).

Both test suites now run with the same PLUGINS=openai-agents value:
  tracing: PLUGINS=openai-agents yarn test:plugins
  llmobs:  PLUGINS=openai-agents yarn test:llmobs:plugins
