+++ packages/datadog-plugin-openai-agents/src/index.js
@@ -3,14 +3,14 @@
 const CompositePlugin = require('../../dd-trace/src/plugins/composite')
 const internalPlugin = require('./internal')
 const clientPlugin = require('./client')
-const llmobsPlugin = require('../../dd-trace/src/llmobs/plugins/openai-agents-core')
+const llmobsPlugin = require('../../dd-trace/src/llmobs/plugins/openai-agents')
 
 class OpenaiAgentsPlugin extends CompositePlugin {
   static id = 'openai-agents'
   static plugins = {
     ...internalPlugin,
-    ...clientPlugin,
     ...llmobsPlugin,
+    ...clientPlugin,
   }
 }
